requirements.txt
biopython
matplotlib
seaborn
rmsd